Add useOpenings hook without pagination
