+++ packages/ui/src/app/pages/WorkingGroups/WorkingGroup/OpeningsTab.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react'
+import React, { useRef } from 'react'
 import styled from 'styled-components'
 
 import { CountBadge } from '@/common/components/CountBadge'
