+++ packages/ui/src/common/components/ContextMenu/ContextMenu.stories.tsx
@@ -1,7 +1,6 @@
 import { Meta, Story } from '@storybook/react'
 import React from 'react'
 
-import { ButtonLink, ButtonPrimary, ButtonSecondary } from '../buttons'
 import { ColumnGapBlock } from '../page/PageContent'
 import { TemplateBlock } from '../storybookParts/previewStyles'
