+++ packages/ui/src/working-groups/components/OpeningsList/OpeningsPagination.tsx
@@ -4,13 +4,14 @@ import { Loading } from '@/common/components/Loading'
 import { Pagination } from '@/common/components/Pagination'
 import { TextBig } from '@/common/components/typography'
 import { OpeningsList } from '@/working-groups/components/OpeningsList/OpeningsList'
-import { useOpenings, UseOpeningsParams } from '@/working-groups/hooks/useOpenings'
+import { UseOpeningsParams } from '@/working-groups/hooks/useOpenings'
+import { useOpeningsPagination } from '@/working-groups/hooks/useOpeningsPagination'
 
-type Props = Omit<UseOpeningsParams, 'page'>
+type Props = UseOpeningsParams
 
 export const OpeningsPagination = (props: Props) => {
   const [page, setPage] = useState(1)
-  const { isLoading, openings, pageCount } = useOpenings({ ...props, page })
+  const { isLoading, openings, pageCount } = useOpeningsPagination({ ...props, page })
 
   if (isLoading) return <Loading />
   if (openings.length === 0) return <TextBig>No openings found</TextBig>
