+++ packages/ui/src/working-groups/components/Roles/RolesList.tsx
@@ -4,7 +4,6 @@ import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { BadgeStatus } from '@/common/components/BadgeStatus'
-import { ButtonLink } from '@/common/components/buttons'
 import { ContextMenu } from '@/common/components/ContextMenu'
 import { List, ListItem } from '@/common/components/List'
 import { TextInlineBig, TokenValue } from '@/common/components/typography'
