+++ packages/ui/src/working-groups/hooks/useOpenings.ts
@@ -1,19 +1,19 @@
 import { useMemo } from 'react'
 
-import { error as logError } from '@/common/logger'
-
-import { useCountWorkingGroupOpeningsQuery, useGetWorkingGroupOpeningsQuery } from '../queries'
-import { asWorkingGroupOpening, WorkingGroupOpeningStatus, WorkingGroupOpeningStatusTypename } from '../types'
-
-export const OPENINGS_PER_PAGE = 5
+import { error } from '@/common/logger'
+import { useGetWorkingGroupOpeningsQuery } from '@/working-groups/queries'
+import {
+  asWorkingGroupOpening,
+  WorkingGroupOpeningStatus,
+  WorkingGroupOpeningStatusTypename,
+} from '@/working-groups/types'
 
 export interface UseOpeningsParams {
   groupId?: string
   statusIn?: WorkingGroupOpeningStatus[]
-  page?: number
 }
 
-const getStatusWhere = (statusIn?: WorkingGroupOpeningStatus[]) => {
+export const getStatusWhere = (statusIn?: WorkingGroupOpeningStatus[]) => {
   if (!statusIn) {
     return
   }
@@ -21,24 +21,13 @@ const getStatusWhere = (statusIn?: WorkingGroupOpeningStatus[]) => {
   return { isTypeOf_in: statusIn.map((status) => WorkingGroupOpeningStatusTypename[status]) }
 }
 
-export const useOpenings = ({ groupId: group_eq, statusIn, page = 1 }: UseOpeningsParams) => {
-  const variables = {
-    limit: OPENINGS_PER_PAGE,
-    offset: (page - 1) * OPENINGS_PER_PAGE,
-    where: {
-      group_eq,
-      status_json: getStatusWhere(statusIn),
-    },
+export const useOpenings = ({ groupId: group_eq, statusIn }: UseOpeningsParams) => {
+  const where = {
+    group_eq,
+    status_json: getStatusWhere(statusIn),
   }
-
-  const { data, loading, error } = useGetWorkingGroupOpeningsQuery({ variables })
-  const { data: countData } = useCountWorkingGroupOpeningsQuery({ variables: variables.where })
-
-  if (error) {
-    logError(error)
-  }
-
-  const totalCount = countData?.workingGroupOpeningsConnection.totalCount ?? 0
+  const { data, loading, error: err } = useGetWorkingGroupOpeningsQuery({ variables: { where } })
+  err && error(err)
 
   const openings = useMemo(() => data?.workingGroupOpenings.map((opening) => asWorkingGroupOpening(opening)) ?? [], [
     loading,
@@ -46,8 +35,7 @@ export const useOpenings = ({ groupId: group_eq, statusIn, page = 1 }: UseOpenin
   ])
 
   return {
-    isLoading: loading,
     openings,
-    pageCount: totalCount && Math.ceil(totalCount / OPENINGS_PER_PAGE),
+    isLoading: loading,
   }
 }
